backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt>=5.3.0
Pillow==10.4.0
drf-yasg==1.21.7
django-cors-headers==4.3.1
gunicorn==21.2.0
python-decouple==3.8
psycopg2-binary>=2.9.9
dj-database-url==2.1.0
whitenoise==6.6.0